refacto mnist
